Couple encrypt and decrypt functions in the same file
