+++ utils/db/github/getToken.ts
@@ -1,5 +1,5 @@
 import executeRequest from "../azuredb";
-import decrypt from "../../encryption/decrypt";
+import {decrypt} from "../../encryption/tokenSalting";
 
 export default async function getUser(user): Promise<any> {
   try {
