+++ utils/encryption/tokenSalting.ts
@@ -12,3 +12,18 @@ export function encrypt(plaintext: string): string {
     }
     return btoa(encrypted);  // encoding result to base64 for better readability
 }
+
+export function decrypt(ciphertext: string): string {
+    if(!process.env.ENCRYPTION_KEY) {
+        throw new Error("Encryption key not found"); 
+    }
+    const key = base64ToUint8Array(process.env.NEXT_PUBLIC_ENCRYPTION_KEY!);
+
+    const decoded = atob(ciphertext);
+    let decrypted = "";
+    
+    for (let i = 0; i < decoded.length; i++) {
+        decrypted += String.fromCharCode(decoded.charCodeAt(i) ^ key[i % key.length]);
+    }
+    return decrypted;
+}
